requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.0.3
fonttools==4.57.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.4
openai==1.30.1
openpyxl==3.1.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pydantic==2.11.3
pydantic_core==2.33.1
PyMuPDF==1.25.5
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.31.0
selenium==4.21.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webdriver-manager==4.0.1
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
WeasyPrint==62.3
psutil==5.9.8
gunicorn==22.0.0
ruff==0.4.8
black==24.4.2
tabulate>=0.9.0
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.8.0
google-api-python-client>=2.0.0
